backend/requirements.txt
flask
flask-cors
pymongo
mongoengine
python-dotenv
redis
solana
solders
gunicorn
